+++ src/shared/components/layouts/MainLayout.tsx
@@ -2,7 +2,7 @@ import NavBar from "../partials/NavBar";
 import HomePage from "../../../pages/HomePage";
 import AboutPage from "../../../pages/AboutPage";
 import FeaturesPage from "../../../pages/FeaturesPage";
-import PricingPage from "../../../pages/PricingPage";
+import PricingPage from "../../../pages/SubscriptionPage";
 import ContactPage from "../../../pages/ContactPage";
 import TestimonialPage from "../../../pages/TestimonialPage";
 import RegistrationPage from "../../../pages/RegistrationPage";
